profile: components for badges and roles
